refactor

+++ Nuget-Publish.ps1
@@ -6,7 +6,10 @@ function Delete-BuildFolders($baseDir)
     foreach($folderToDelete in @('bin', 'obj'))
     {
         $absoluteFolderPath = [System.IO.Path]::Combine($baseDir, $folderToDelete)
-        Remove-Item $absoluteFolderPath -Force -Recurse
+        if(Test-Path $absoluteFolderPath)
+        {
+            Remove-Item $absoluteFolderPath -Force -Recurse
+        }
     }
 }
 
@@ -14,4 +17,5 @@ foreach($projectName in $projectNames)
 {
     $projectDirectory = [System.IO.Path]::Combine($PSScriptRoot, $projectName)
     Delete-BuildFolders $projectDirectory
+    # http://stackoverflow.com/questions/4677222/powershell-script-to-build-visual-studio-project
 }
